network intrusions is a python notebook
network_intrusions is a jupyter notebook where we explore a Kaggle data set which collects normal and malicious connections made to a server.  We explore this data set and model it using GradientBoostingRegressor and RandomForestRegressor.  This model was chosen for this exploration given its allowance for NaN values.  An accuracy of 0.89 was achieved which is within the acceptance range without risking overfitting of the model.