Jupyter notebook of Credit Fraud
Here we take a look at another kaggle dataset collected from numerous financial institutions before and after major fraud incidents. We are running an unsupervised model to attain the best and most predictive results from this dataset.  The Pycaret was used to be specific. this model came out with an accuracy score of 0.98.